requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
google-cloud-secret-manager>=2.18.0
google-api-python-client>=2.100.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
cloud-sql-python-connector>=1.11.0
psycopg2-binary==2.9.9
alembic==1.13.1
firebase-admin>=6.2.0
google-cloud-firestore>=2.11.0
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0
aiofiles>=23.2.1

# MCP Management System dependencies
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
httpx>=0.24.0
PyJWT>=2.8.0# Force rebuild Mon Aug 11 12:44:08 CDT 2025
